doc: Add an explanation about the log format
I noticed that, albeit old, issue #21 talks about pino-colada swallowing
errors. This simply isn't true in the latest releases of pino-colada.
Therefore, I have added some documentation to the readme that explains
the format and where each of the displayed items is coming from within
the log messages. This closes #21

Signed-off-by: Colton Wolkins (Indicio work address) <colton@indicio.tech>